requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
drf-yasg==1.21.7
ijson==3.3.0
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.2
setuptools==74.0.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
uritemplate==4.1.1
